requirements.txt
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
filelock==3.20.0
fsspec==2025.10.0
generativeai==0.0.1
google-auth==2.43.0
google-genai==1.52.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.7
idna==3.11
jiter==0.12.0
openai==2.8.1
packaging==25.0
playwright==1.56.0
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
shellingham==1.5.4
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
